+++ ent/fuzzy/fuzzy_test.go
@@ -25,10 +25,10 @@ func TestDist(t *testing.T) {
 		{"Pomatomus", "Pom-tomus", 1},
 		{"Pomatomus", "Pomщtomus", 1},
 		// ed = 3, too big
-		{"sitting", "kitten", -1},
+		{"sitting", "kitten", 3},
 		// words are too small
-		{"Pom atomus", "Poma tomus", -1},
-		{"Acacia mal", "Acacia may", -1},
+		{"Pom atomus", "Poma tomus", 2},
+		{"Acacia mal", "Acacia may", 1},
 		// differnt number of words is not covered yet
 		{"Pomatomus", "Poma  tomus", 2},
 		// edge cases that should not happen